% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean Rust binaries and package cache.}
\usage{
clean(path = ".")
}
\arguments{
\item{path}{[ string ] Path to the package root.}
}
\description{
Removes Rust binaries (such as \code{.dll}/\code{.so} libraries), C wrapper object files,
invokes \verb{cargo clean} to reset cargo target directory
(found by default at \verb{pkg_root/src/rust/target/}).
Useful when Rust code should be recompiled from scratch.
}
