% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rextendr_document.R
\name{document}
\alias{document}
\title{Compile Rust code and generate package documentation.}
\usage{
document(pkg = ".", quiet = FALSE, roclets = NULL)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[devtools:as.package]{as.package()}} for more information.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}

\item{roclets}{Character vector of roclet names to use with package.
The default, \code{NULL}, uses the roxygen \code{roclets} option,
which defaults to \code{c("collate", "namespace", "rd")}.}
}
\value{
No return value, called for side effects.
}
\description{
As of \code{rextendr} 0.4.0, this function is no longer strictly necessary.
Packages created with \code{\link[=use_extendr]{use_extendr()}} now include a \code{document} binary that
generates \code{R/extendr-wrappers.R} as part of the normal \verb{cargo build} step,
so \code{\link[devtools:document]{devtools::document()}} works directly without any \code{rextendr}-specific
pre-processing. \code{rextendr::document()} is retained for backwards
compatibility.
}
\details{
\code{rextendr::document()} updates the package documentation for an R package
that uses \code{extendr} code. It is a wrapper for \code{\link[devtools:document]{devtools::document()}}.
}
