% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filetracker_roclet.R
\name{filetracker_roclet}
\alias{filetracker_roclet}
\title{Create roclet that tracks Rust source files changes.}
\usage{
filetracker_roclet()
}
\value{
A roclet.
}
\description{
Create roclet that tracks Rust source files changes.
}
