% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_rust_source.R
\name{get_rust_files}
\alias{get_rust_files}
\title{Obtains paths to the Rust source files.}
\usage{
get_rust_files(path = ".")
}
\arguments{
\item{path}{Path from which package root is looked up.}
}
\value{
A vector of file paths (can be empty).
}
\description{
Enumerates all Rust files, changes to which should trigger re-compilation.
This inclides \verb{*.rs} and \code{Cargo.toml}.
}
\keyword{internal}
