% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inf_dev_extendr_used}
\alias{inf_dev_extendr_used}
\title{Inform the user that a development version of \code{extendr} is being used.}
\usage{
inf_dev_extendr_used()
}
\description{
This function returns a string that should be used inside of a \code{cli} function.
See \code{validate_extendr_features()} for an example.
}
\keyword{internal}
