% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_extendr.R
\name{make_example_wrappers}
\alias{make_example_wrappers}
\title{Creates example R wrappers for Rust functions.}
\usage{
make_example_wrappers(
  pkg_name,
  outfile,
  path = ".",
  extra_items = NULL,
  quiet = FALSE
)
}
\arguments{
\item{pkg_name}{The name of the package.}

\item{outfile}{Determines where to write wrapper code.}

\item{path}{Path from which package root is looked up. Used for pretty message formatting.}

\item{extra_items}{Character vector or \code{NULL} containing additional wrappers.
Should be valid R code.}

\item{quiet}{Logical scalar indicating whether the output should be quiet (\code{TRUE})
or verbose (\code{FALSE}).}
}
\description{
Writes simple R wrappers, without which initial compilation of the
package is impossible. Does not require a compiled library.
Can be used as a fallback.
}
\keyword{internal}
