% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_extendr.R
\name{make_wrappers}
\alias{make_wrappers}
\title{Creates R wrappers for Rust functions.}
\usage{
make_wrappers(
  module_name,
  package_name,
  outfile,
  path = ".",
  use_symbols = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{module_name}{The name of the Rust module. Can be the same as \code{package_name}}

\item{package_name}{The name of the package.}

\item{outfile}{Determines where to write wrapper code.}

\item{path}{Path from which package root is looked up. Used for message formatting.}

\item{use_symbols}{Logical, indicating wether to add additonal symbol information to
the generated wrappers. Default (\code{FALSE}) is used when making wrappers for the package,
while \code{TRUE} is used to make wrappers for dynamically generated libraries using
\code{\link{rust_source}}, \code{\link{rust_function}}, etc.}

\item{quiet}{Logical scalar indicating whether the output should be quiet (\code{TRUE})
or verbose (\code{FALSE}).}
}
\description{
Invokes \verb{wrap__make_\{module_name\}_wrappers} exported from
the Rust library and writes obtained R wrappers to the \code{outfile}.
}
\keyword{internal}
