% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cargo_metadata.R
\name{read_cargo_metadata}
\alias{read_cargo_metadata}
\title{Retrieve metadata for packages and workspaces}
\usage{
read_cargo_metadata(path = ".")
}
\arguments{
\item{path}{character scalar, the R package directory}
}
\value{
\code{list}, including the following elements:
\itemize{
\item "packages"
\item "workspace_members"
\item "workspace_default_members"
\item "resolve"
\item "target_directory"
\item "version"
\item "workspace_root"
\item "metadata"
}
}
\description{
Retrieve metadata for packages and workspaces
}
\details{
For more details, see
\href{https://doc.rust-lang.org/cargo/commands/cargo-metadata.html}{Cargo docs}
for \code{cargo-metadata}. See especially "JSON Format" to get a sense of what you
can expect to find in the returned list.
}
\examples{
\dontrun{
read_cargo_metadata()
}

}
