% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_extendr.R
\name{register_extendr}
\alias{register_extendr}
\title{Register the extendr module of a package with R}
\usage{
register_extendr(path = ".", quiet = FALSE, force = FALSE, compile = NA)
}
\arguments{
\item{path}{Path from which package root is looked up.}

\item{quiet}{Logical indicating whether any progress messages should be
generated or not.}

\item{force}{Logical indicating whether to force regenerating
\code{R/extendr-wrappers.R} even when it doesn't seem to need updated. (By
default, generation is skipped when it's newer than the DLL).}

\item{compile}{Logical indicating whether to recompile DLLs:
\describe{
\item{\code{TRUE}}{always recompiles}
\item{\code{NA}}{recompiles if needed (i.e., any source files or manifest file are newer than the DLL)}
\item{\code{FALSE}}{never recompiles}
}}
}
\value{
(Invisibly) Path to the file containing generated wrappers.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated because we now rely on a small Rust binary to
generate wrappers, which is called during the package build process.
}
\seealso{
\code{\link[=document]{document()}}
}
