% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration_roclet.R
\name{registration_roclet}
\alias{registration_roclet}
\title{Create roclet that registers \code{extendr} package
and updates wrappers.}
\usage{
registration_roclet()
}
\value{
A roclet.
}
\description{
Create roclet that registers \code{extendr} package
and updates wrappers.
}
