% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rextendr.R
\docType{package}
\name{rextendr}
\alias{rextendr-package}
\alias{rextendr}
\title{Call Rust code from R using the 'extendr' Crate}
\description{
The rextendr package implements functions to interface with Rust code from R.
See \code{\link[=rust_source]{rust_source()}} for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://extendr.rs/rextendr/}
  \item \url{https://github.com/extendr/rextendr}
  \item Report bugs at \url{https://github.com/extendr/rextendr/issues}
}

}
\author{
\strong{Maintainer}: Kenneth Vernon \email{kenneth.b.vernon@gmail.com} (\href{https://orcid.org/0000-0003-0098-5092}{ORCID})

Authors:
\itemize{
  \item Kenneth Vernon \email{kenneth.b.vernon@gmail.com} (\href{https://orcid.org/0000-0003-0098-5092}{ORCID})
  \item Claus O. Wilke \email{wilke@austin.utexas.edu} (\href{https://orcid.org/0000-0002-7470-9261}{ORCID})
  \item Andy Thomason \email{andy@andythomason.com}
  \item Mossa M. Reimert \email{mossa@sund.ku.dk}
  \item Ilia Kosenkov \email{ilia.kosenkov@outlook.com} (\href{https://orcid.org/0000-0001-5563-7840}{ORCID})
  \item Malcolm Barrett \email{malcolmbarrett@gmail.com} (\href{https://orcid.org/0000-0003-0299-5825}{ORCID})
  \item Josiah Parry \email{josiah.parry@gmail.con} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})
}

Other contributors:
\itemize{
  \item Alberson Miranda \email{albersonmiranda@hotmail.com} (\href{https://orcid.org/0000-0001-9252-4175}{ORCID}) [contributor]
}

}
\keyword{internal}
