% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust_sitrep.R
\name{rust_sitrep}
\alias{rust_sitrep}
\title{Report on Rust infrastructure}
\usage{
rust_sitrep()
}
\value{
Nothing
}
\description{
Prints out a detailed report on the state of Rust infrastructure on the host machine.
}
