% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{rust_source}
\alias{rust_source}
\alias{rust_function}
\alias{extendr_options}
\alias{print.extendr_opts}
\title{Compile Rust code and call from R}
\usage{
rust_source(
  file = NULL,
  code = NULL,
  env = parent.frame(),
  echo = FALSE,
  quiet = FALSE,
  opts = NULL,
  ...
)

rust_function(
  code,
  extendr_fn_options = NULL,
  env = parent.frame(),
  echo = FALSE,
  quiet = FALSE,
  opts = NULL,
  ...
)

extendr_options(
  cache_build = TRUE,
  dependencies = NULL,
  extendr_deps = NULL,
  features = NULL,
  generate_module_macro = TRUE,
  module_name = "rextendr",
  patch.crates_io = getOption("rextendr.patch.crates_io"),
  profile = c("dev", "release", "perf"),
  toolchain = getOption("rextendr.toolchain"),
  use_dev_extendr = FALSE,
  use_extendr_api = TRUE,
  use_rtools = TRUE
)

\method{print}{extendr_opts}(x, ...)
}
\arguments{
\item{file}{character scalar, input rust file to source.}

\item{code}{character scalar, input rust code to be used instead of \code{file}.}

\item{env}{environment, the R environment in which the wrapping functions
will be defined. Default is \code{parent.frame()}.}

\item{echo}{logical scalar, whether to print standard output and errors of
\code{cargo} commands to the console. Default is \code{FALSE}.}

\item{quiet}{logical scalar, whether to print \code{cli} errors and warnings.
Default is \code{FALSE}.}

\item{opts}{\code{extendr_opts} list, set using \code{extendr_options()}. Default is
\code{NULL}.}

\item{...}{user supplied extendr options to be injected into the
\code{extendr_opts} list (for backwards compatibility).}

\item{extendr_fn_options}{A list of extendr function options that are
inserted into the \verb{#[extendr(...)]} attribute}

\item{cache_build}{logical scalar, whether builds should be cached between
calls to \code{\link[=rust_source]{rust_source()}}.}

\item{dependencies}{character vector, dependencies to be added to \code{Cargo.toml}.}

\item{extendr_deps}{named list, versions of \verb{extendr-*} crates. Defaults to
\code{rextendr.extendr_deps} option (\code{list(`extendr-api` = "*")}) if
\code{use_dev_extendr} is not \code{TRUE}, otherwise, uses
\code{rextendr.extendr_dev_deps} option (\code{list(`extendr-api` = list(git =
"https://github.com/extendr/extendr")}).}

\item{features}{character vector, \code{extendr-api} features that should be enabled.
Supported values are \code{"ndarray"}, \code{"faer"}, \code{"either"}, \code{"num-complex"}, \code{"serde"}, and \code{"graphics"}.
Unknown features will produce a warning if \code{quiet} is not \code{TRUE}.}

\item{generate_module_macro}{logical scalar, whether the Rust module
macro should be automatically generated from the code. Default is \code{TRUE}.
Ignored for Rust source provided via \code{file}. The macro generation is done
with \code{\link[=make_module_macro]{make_module_macro()}} and it may fail in complex cases. If something
doesn't work, try calling \code{\link[=make_module_macro]{make_module_macro()}} on your code to see whether
the generated macro code has issues.}

\item{module_name}{character scalar, name of the module defined in the Rust source via
\verb{extendr_module!}. Default is \code{"rextendr"}. If \code{generate_module_macro} is \code{FALSE}
or if \code{file} is specified, should \emph{match exactly} the name of the module defined in the source.}

\item{patch.crates_io}{character vector, patch statements for crates.io to
be added to \code{Cargo.toml}.}

\item{profile}{character scalar, Rust profile. Can be either \code{"dev"}, \code{"release"} or \code{"perf"}.
The default, \code{"dev"}, compiles faster but produces slower code.}

\item{toolchain}{character scalar, Rust toolchain. The default, \code{NULL}, compiles with the
system default toolchain. Accepts valid Rust toolchain qualifiers,
such as \code{"nightly"}, or (on Windows) \code{"stable-msvc"}.}

\item{use_dev_extendr}{logical scalar, whether to use development version of
\code{extendr}. Has no effect if \code{extendr_deps} are set.}

\item{use_extendr_api}{logical scalar, whether \verb{use extendr_api::prelude::*;}
should be added at the top of the Rust source provided via \code{code}. Default
is \code{TRUE}. Ignored for Rust source provided via \code{file}.}

\item{use_rtools}{logical scalar, whether to append the path to Rtools to the
\code{PATH} variable on Windows using the \code{RTOOLS4X_HOME} environment variable
(if it is set). The appended path depends on the process architecture. Does
nothing on other platforms.}

\item{x}{an \code{extendr_opts} list}
}
\value{
For \code{rust_source()} and \code{rust_function()}, the result from
\code{\link[=dyn.load]{dyn.load()}}, which is an object of class \code{DLLInfo}. See \code{\link[=getLoadedDLLs]{getLoadedDLLs()}}
for more details. For \code{extendr_options()}, an \code{extendr_opts} list.
}
\description{
\code{\link[=rust_source]{rust_source()}} compiles and loads a single Rust file for use in R.
\code{\link[=rust_function]{rust_function()}} compiles and loads a single Rust function for use in R.
\code{\link[=extendr_options]{extendr_options()}} is a helper function to make it easier to pass additional
options when sourcing Rust code. It also proivdes defaults for each option
and does additional type checking.
}
\examples{
\dontrun{
# creating a single rust function
rust_function("fn add(a:f64, b:f64) -> f64 { a + b }")
add(2.5, 4.7)

# creating multiple rust functions at once
code <- r"(
#[extendr]
fn hello() -> &'static str {
    "Hello, world!"
}

#[extendr]
fn test( a: &str, b: i64) {
    rprintln!("Data sent to Rust: {}, {}", a, b);
}
)"

rust_source(code = code)
hello()
test("a string", 42)

# use case with an external dependency: a function that converts
# markdown text to html, using the `pulldown_cmark` crate.
code <- r"(
  use pulldown_cmark::{Parser, Options, html};

  #[extendr]
  fn md_to_html(input: &str) -> String {
    let mut options = Options::empty();
    options.insert(Options::ENABLE_TABLES);
    let parser = Parser::new_ext(input, options);
    let mut output = String::new();
    html::push_html(&mut output, parser);
    output
  }
)"

rust_source(
  code = code,
  opts = extendr_options(
    dependencies = list(`pulldown-cmark` = "0.8")
  )
)

md_text <- "# The story of the fox
The quick brown fox **jumps over** the lazy dog.
The quick *brown fox* jumps over the lazy dog."

md_to_html(md_text)

# see default options
extendr_options()
}
}
