% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_save_all.R
\name{try_save_all}
\alias{try_save_all}
\title{Try to save open files if \pkg{rextendr} is called from an IDE.}
\usage{
try_save_all(quiet = FALSE)
}
\arguments{
\item{quiet}{Logical scalar indicating whether the output should be quiet (\code{TRUE})
or verbose (\code{FALSE}).}
}
\description{
Uses rstudio API (if available) to save modfied files.
Improves package development experience within RStudio.
}
\keyword{internal}
