% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_i}
\alias{ui_i}
\title{Formats text as an information message.}
\usage{
ui_i(text = "")
}
\arguments{
\item{text}{String to format.}
}
\description{
Prepends \code{text} with cyan info sign (\code{i}). Performs \code{\link[glue:glue]{glue::glue()}} interpolation.
}
\keyword{internal}
