% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_q}
\alias{ui_q}
\title{Formats text as a question message.}
\usage{
ui_q(text = "")
}
\arguments{
\item{text}{String to format.}
}
\description{
Prepends \code{text} with yellow question mark (\verb{?}). Performs \code{\link[glue:glue]{glue::glue()}} interpolation.
}
\keyword{internal}
