% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_throw}
\alias{ui_throw}
\title{Throws an error with formatted message.}
\usage{
ui_throw(message, details = character(0))
}
\arguments{
\item{message}{The primary error message. Mandatory.}

\item{details}{An optional character vector of error detais.
can be formatted with \verb{ui_*} helper functions.}
}
\description{
Creates a styled error message that is then thrown
using \code{\link[=stop]{stop()}}.
}
\examples{
\dontrun{
ui_throw(
  "Something bad has happened!",
  c(
    ui_x("This thing happened."),
    ui_x("That thing happened."),
    ui_q("Are you sure you did it right?")
  )
)
# Error: Something bad has happened!
# x This thing happened.
# x That thing happened.
# ? Are you sure you did it right?
}
}
\keyword{internal}
