% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_v}
\alias{ui_v}
\title{Formats text as a success message.}
\usage{
ui_v(text = "")
}
\arguments{
\item{text}{String to format.}
}
\description{
Prepends \code{text} with green check mark (\code{v}). Performs \code{\link[glue:glue]{glue::glue()}} interpolation.
}
\keyword{internal}
