% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_w}
\alias{ui_w}
\title{Formats text as a warning message.}
\usage{
ui_w(text = "")
}
\arguments{
\item{text}{String to format.}
}
\description{
Prepends \code{text} with yellow exclamation mark (\code{!}). Performs \code{\link[glue:glue]{glue::glue()}} interpolation.
}
\keyword{internals}
