% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui_x}
\alias{ui_x}
\title{Formats text as an error message.}
\usage{
ui_x(text = "")
}
\arguments{
\item{text}{String to format.}
}
\description{
Prepends \code{text} with red cross (\code{x}). Performs \code{\link[glue:glue]{glue::glue()}} interpolation.
}
\keyword{internal}
