% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_extendr.R
\name{update_extendr}
\alias{update_extendr}
\title{Update extendr scaffolding}
\usage{
update_extendr(
  path = ".",
  crate_name = NULL,
  lib_name = NULL,
  revendor = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{path}{File path to the package for which to generate wrapper code.}

\item{crate_name}{String that is used as the name of the Rust crate,
specifically \verb{[package] name} in \code{Cargon.toml}. If \code{NULL} (default),
sanitized R package name is used instead.}

\item{lib_name}{String that is used as the name of the Rust library,
specifically \verb{[lib] name} in \code{Cargo.toml}. If \code{NULL} (default), sanitized R
package name is used instead.}

\item{revendor}{boolean scalar, whether to clear vendor files and re-run
\code{rextendr::vendor_pkgs()} (default is \code{TRUE}).}

\item{quiet}{Logical indicating whether any progress messages should be
generated or not.}
}
\value{
a logical scalar indicating whether updating was successful
}
\description{
When a new extendr or rextendr release requires changes to scaffolding, this
function helps update those files to the new specification.
}
\details{
Unfortunately, this process cannot be fully automated, so
information is also printed to the console explaining what needs to be
updated by hand. Usually, this will be accompanied by a more detailed blog
post explaining changes.
\subsection{Current list of updated files:}{
\itemize{
\item \code{src/entrypoint.c}
\item \code{src/Makevars.in}
\item \code{src/Makevars.win.in}
\item \code{cleanup}
\item \code{cleanup.win}
\item \code{src/rust/document.rs}
\item \code{tools/msrv.R}
\item \code{tools/config.R}
\item \code{configure}
\item \code{configure.win}
}
}

\subsection{Additionally updated when \code{revendor = TRUE}:}{
\itemize{
\item \verb{src/rust/vendor/}
\item \code{src/rust/vendor.tar.xz}
\item \code{src/rust/vendor-config.toml}
}
}
}
