% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_scaffold.R
\name{update_scaffold}
\alias{update_scaffold}
\title{Update extendr scaffolding}
\usage{
update_scaffold(path = ".", crate_name = NULL, lib_name = NULL, quiet = FALSE)
}
\arguments{
\item{path}{File path to the package for which to generate wrapper code.}

\item{crate_name}{String that is used as the name of the Rust crate,
specifically \verb{[package] name} in \code{Cargon.toml}. If \code{NULL} (default),
sanitized R package name is used instead.}

\item{lib_name}{String that is used as the name of the Rust library,
specifically \verb{[lib] name} in \code{Cargo.toml}. If \code{NULL} (default), sanitized R
package name is used instead.}

\item{quiet}{Logical indicating whether any progress messages should be
generated or not.}
}
\value{
a logical scalar indicating whether scaffold updating was successful
}
\description{
When a new version of extendr or rextendr is released, this function updates
relevant scaffolding files to the new specification and removes any artifacts
that may affect dependency resolution, with the important exception of
Cargo.lock and Cargo.toml, which are left as is.
}
\details{
This is only \emph{part} of the update process! In fact, this just does
some basic file housekeeping, so that the developer can focus on the
important part, namely updating Cargo.lock and Cargo.toml to use the newest
version of extendr-api. Note that updating to the newest extendr version
may require some manual dependency resolution.

The following files are rewritten to the current specification:
\itemize{
\item \code{src/entrypoint.c}
\item \code{src/Makevars.in}
\item \code{src/Makevars.win.in}
\item \code{cleanup}
\item \code{cleanup.win}
\item \code{src/rust/document.rs}
\item \code{tools/msrv.R}
\item \code{tools/config.R}
\item \code{configure}
\item \code{configure.win}
}

The following build artifacts and vendoring files are removed if present:
\itemize{
\item \verb{src/rust/vendor/}
\item \code{src/rust/vendor.tar.xz}
\item \code{src/rust/vendor-config.toml}
\item \verb{src/rust/.cargo/}
\item \verb{src/rust/target/}
}

After clearing these files and folders, \code{update_scaffold()} will print a
message that explains what to do next (provided \code{quiet = FALSE}). Usually,
this will be accompanied by a more detailed blog post explaining the update
process.
}
