% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendor.R
\name{use_cargo_vendor}
\alias{use_cargo_vendor}
\alias{vendor_pkgs}
\title{Vendor Cargo Dependencies}
\usage{
use_cargo_vendor(path = ".", quiet = FALSE, overwrite = NULL, lib_name = NULL)

vendor_pkgs(path = ".", quiet = FALSE)
}
\arguments{
\item{path}{File path to the package for which to generate wrapper code.}

\item{quiet}{Logical indicating whether any progress messages should be
generated or not.}

\item{overwrite}{Logical scalar or \code{NULL} indicating whether the files in the \code{path} should be overwritten.
If \code{NULL} (default), the function will ask the user whether each file should
be overwritten in an interactive session or do nothing in a non-interactive session.
If \code{FALSE} and each file already exists, the function will do nothing.
If \code{TRUE}, all files will be overwritten.}

\item{lib_name}{String that is used as the name of the Rust library.
If \code{NULL}, sanitized R package name is used instead.}
}
\value{
\itemize{
\item \code{vendor_pkgs()} returns a data.frame with two columns \code{crate} and \code{version}
\item \code{use_cargo_vendor()} returns \code{NULL} and used solely for its side effects
}
}
\description{
Creates the necessary infrastructure to vendor
Rust dependencies with an R package.
}
\details{
CRAN requires that any R package that uses Rust must also include
its dependencies in the package itself. This is to ensure that a
package can be installed in an offline manner.

\code{use_cargo_vendor()} creates the scaffolding necessary to vendor dependencies. These
are a modified \code{Makevars}, \code{Makevars.win}, and \code{vendor-config.toml} files. In
addition, it will add requisite lines to your \code{.gitignore} and \code{.Rbuildignore}
if usethis is available.

\code{vendor_pkgs()} is used to package the dependencies. It compresses the \code{vendor}
directory into a single \code{vendor.tar.xz} which will be shipped with package itself.
If you have modified your dependencies, you will need need to repackage
}
\examples{

if (interactive()) {
  use_cargo_vendor()
  vendor_pkgs()
}
}
