% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-version.R
\name{use_version}
\alias{use_version}
\alias{sync_version}
\title{Modify Package Version}
\usage{
use_version(which = c("major", "minor", "patch"))

sync_version(path = ".")
}
\description{
Modify Package Version
}
