% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_quiet}
\alias{with_quiet}
\title{Silence \code{{cli}} output}
\usage{
with_quiet(quiet, env = rlang::caller_env())
}
\description{
Use for functions that use cli output that should optionally be suppressed.
}
\examples{

if (interactive()) {
hello_rust <- function(..., quiet = FALSE) {
  with_quiet(quiet)
  cli::cli_alert_info("This should be silenced when {.code quiet = TRUE}")
}

hello_rust()
hello_rust(quiet = TRUE)
}
}
\keyword{internal}
