% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_extendr.R
\name{write_namespace}
\alias{write_namespace}
\title{Appends \code{useDynLib} to the \code{NAMESPACE} file.}
\usage{
write_namespace(pkg_name, path = ".", quiet = FALSE)
}
\arguments{
\item{pkg_name}{Name of the package (should be identical to the library name without extension).}

\item{path}{Path from which package root is looked up.}

\item{quiet}{Logical scalar indicating whether the output should be quiet (\code{TRUE})
or verbose (\code{FALSE}).}
}
\description{
If \code{NAMESPACE} is missing \code{useDynLib} entry, loading package will not
load the native library containing Rust code. This prevents any \code{\link{.Call}}s,
which makes wrapper generation impossible.
\code{NAMESPACE} contents are searched for \code{useDynLib} entry, and if it is not found,
it is appended.
This normally happens only once for each developed package -- immediately after
package has been created and its \code{NAMESPACE} is empty.
}
\keyword{internal}
