% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rextendr_document.R
\name{document}
\alias{document}
\title{Compile Rust code and generate package documentation.}
\usage{
document(pkg = ".", quiet = getOption("usethis.quiet", FALSE), roclets = NULL)
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See \code{\link[devtools:as.package]{as.package()}} for more information.}

\item{quiet}{if \code{TRUE} suppresses output from this function.}

\item{roclets}{Character vector of roclet names to use with package.
The default, \code{NULL}, uses the roxygen \code{roclets} option,
which defaults to \code{c("collate", "namespace", "rd")}.}
}
\value{
No return value, called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\details{
\code{rextendr::document()} was a wrapper for \code{\link[devtools:document]{devtools::document()}} to ensure
that Rust code is recompiled (when necessary) and that up-to-date R wrappers
are generated before re-generating the package documentation. However, as of
pkgdown 1.4.0, \code{devtools::document()} detects the changes on Rust code, so
the users can rely on it directly.
}
