% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rextendr_document.R
\name{document}
\alias{document}
\title{This function is a wrapper for \code{devtools::document}.
It executes \code{extendr}-specific routine before calling \code{devtools::document},
ensuring Rust code is recompiled when necessary and appropriate R wrappers are generated.}
\usage{
document(pkg = ".", quiet = FALSE, roclets = NULL)
}
\arguments{
\item{pkg}{String scalar pointing to the package root.}

\item{quiet}{Logical scalar indicating whether the output should be quiet (\code{TRUE})
or verbose (\code{FALSE}).}

\item{roclets}{Additional roclets from \code{roxygen2} that are passed into
\code{roxygen2::document}.}
}
\description{
This function is a wrapper for \code{devtools::document}.
It executes \code{extendr}-specific routine before calling \code{devtools::document},
ensuring Rust code is recompiled when necessary and appropriate R wrappers are generated.
}
