% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_engine.R
\name{eng_extendr}
\alias{eng_extendr}
\title{Knitr engine}
\usage{
eng_extendr(options)
}
\arguments{
\item{options}{A list of chunk options.}
}
\value{
A character string representing the engine output.
}
\description{
Knitr engine
}
