% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_engine.R
\name{eng_extendr}
\alias{eng_extendr}
\alias{eng_extendrfuns}
\title{Knitr engines}
\usage{
eng_extendr(options)

eng_extendrfuns(options)
}
\arguments{
\item{options}{A list of chunk options.}
}
\value{
A character string representing the engine output.
}
\description{
Two knitr engines that enable code chunks of type \code{extendr} (individual Rust
statements to be evaluated) and \code{extendrfuns} (Rust functions that will be
exported to R).
}
