% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_engine.R
\name{eng_extendr}
\alias{eng_extendr}
\alias{eng_extendrsrc}
\title{Knitr engines}
\usage{
eng_extendr(options)

eng_extendrsrc(options)
}
\arguments{
\item{options}{A list of chunk options.}
}
\value{
A character string representing the engine output.
}
\description{
Two knitr engines that enable code chunks of type \code{extendr} (individual Rust
statements to be evaluated via \code{\link[=rust_eval]{rust_eval()}}) and \code{extendrsrc} (Rust functions
or classes that will be exported to R via \code{\link[=rust_source]{rust_source()}}).
}
