% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_libR_bindings.R
\name{install_libR_bindings}
\alias{install_libR_bindings}
\title{Install libR bindings for Rust}
\usage{
install_libR_bindings(
  version = "*",
  force = FALSE,
  quiet = FALSE,
  patch.crates_io = NULL
)
}
\arguments{
\item{version}{libR-sys version, provided as a Rust version string. \code{"*"} will install the
latest available version on crates.io.}

\item{force}{Logical indicating whether install should be forced
even if bindings have already been installed previously.}

\item{quiet}{Logical indicating whether compile output should be generated or not.}

\item{patch.crates_io}{Character vector of patch statements for crates.io to
be added to the \code{Cargo.toml} file.}
}
\value{
Integer error code as returned by \code{\link[=system2]{system2()}}. A value of \code{0L} indicates success.
}
\description{
Call \code{\link[=install_libR_bindings]{install_libR_bindings()}} once after installing rextendr to locally prebuild and install
libR bindings for Rust. This is not required for any other parts of the package to function
properly, but it will speed up subsequent calls to \code{\link[=rust_source]{rust_source()}} as the bindings don't have
to be regenerated over and over.
}
