% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_extendr.R
\name{make_wrappers_externally}
\alias{make_wrappers_externally}
\title{Creates R wrappers for Rust functions.}
\usage{
make_wrappers_externally(
  module_name,
  package_name,
  outfile,
  path,
  use_symbols = FALSE,
  quiet = FALSE,
  compile = NA
)
}
\arguments{
\item{module_name}{The name of the Rust module. Can be the same as \code{package_name}}

\item{package_name}{The name of the package.}

\item{outfile}{Determines where to write wrapper code.}

\item{path}{Path from which package root is looked up. Used for message formatting.}

\item{use_symbols}{Logical, indicating wether to add additonal symbol information to
the generated wrappers. Default (\code{FALSE}) is used when making wrappers for the package,
while \code{TRUE} is used to make wrappers for dynamically generated libraries using
\code{\link{rust_source}}, \code{\link{rust_function}}, etc.}

\item{quiet}{Logical scalar indicating whether the output should be quiet (\code{TRUE})
or verbose (\code{FALSE}).}

\item{compile}{Logical indicating whether to recompile DLLs:
\describe{
\item{\code{TRUE}}{always recompiles}
\item{\code{NA}}{recompiles if needed (i.e., any source files or manifest file are newer than the DLL)}
\item{\code{FALSE}}{never recompiles}
}}
}
\description{
Does the same as \code{\link{make_wrappers}}, but out of process.
}
\keyword{internal}
