% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_rust_source.R
\name{needs_compilation}
\alias{needs_compilation}
\title{Checks if re-compilation is needed.
Tracks changes in Rust source files (\verb{*.rs}) and \code{Cargo.toml}.}
\usage{
needs_compilation(path = ".", quiet = FALSE)
}
\arguments{
\item{path}{Package root.}

\item{quiet}{Logical scalar indicating wether the output should be quiet (\code{TRUE})
or verbose (\code{FALSE}).}
}
\value{
Logical \code{TRUE} if Rust source has been modified, \code{FALSE} otherwise.
}
\description{
Checks if re-compilation is needed.
Tracks changes in Rust source files (\verb{*.rs}) and \code{Cargo.toml}.
}
