% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rextendr.R
\docType{package}
\name{rextendr}
\alias{rextendr}
\title{Call Rust code from R using the 'extendr' Crate}
\description{
The rextendr package implements functions to interface with Rust code from R.
See \code{\link[=rust_source]{rust_source()}} for details.
}
