% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{rust_eval}
\alias{rust_eval}
\title{Evaluate Rust code}
\usage{
rust_eval(code, ...)
}
\arguments{
\item{code}{Input rust code.}

\item{...}{Other parameters handed off to \code{\link[=rust_function]{rust_function()}}.}
}
\description{
Evaluate Rust code.
}
\examples{
\dontrun{
# Rust code without return value, called only for its side effects
rust_eval(
  code = 'rprintln!("hello from Rust!");'
)

# Rust code with return value; note the `.into()` at the end
rust_eval(
  code = '
    let x = 5;
    let y = 7;
    let z = x * y;
    z.into() // cast z into an Robj and return
 '
)
}
}
