% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{rust_source}
\alias{rust_source}
\title{Compile Rust code}
\usage{
rust_source(
  file,
  code = NULL,
  env = parent.frame(),
  clean = TRUE,
  quiet = TRUE
)
}
\description{
\code{\link[=rust_source]{rust_source()}} compiles and loads a single Rust file for use in R.
}
