% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{rust_source}
\alias{rust_source}
\title{Compile Rust code}
\usage{
rust_source(
  file,
  code = NULL,
  dependencies = NULL,
  patch.crates_io = NULL,
  env = parent.frame(),
  cache_build = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{file}{Input rust file to source.}

\item{code}{Input rust code, to be used instead of \code{file}.}

\item{dependencies}{Character vector of dependencies lines to be added to the
\code{Cargo.toml} file.}

\item{patch.crates_io}{Character vector of patch statements for crates.io to
be added to the \code{Cargo.toml} file.}

\item{env}{The R environment in which the wrapping functions will be defined.}

\item{cache_build}{Logical indicating whether builds should be cached between
calls to \code{\link[=rust_source]{rust_source()}}.}

\item{quiet}{Logical indicating whether compile output should be generated or not.}
}
\description{
\code{\link[=rust_source]{rust_source()}} compiles and loads a single Rust file for use in R.
}
