% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_scaffold.R
\name{update_scaffold}
\alias{update_scaffold}
\title{Update extendr scaffolding}
\usage{
update_scaffold(path = ".", crate_name = NULL, lib_name = NULL, quiet = FALSE)
}
\arguments{
\item{path}{File path to the package for which to generate wrapper code.}

\item{crate_name}{String that is used as the name of the Rust crate,
specifically \verb{[package] name} in \code{Cargon.toml}. If \code{NULL} (default),
sanitized R package name is used instead.}

\item{lib_name}{String that is used as the name of the Rust library,
specifically \verb{[lib] name} in \code{Cargo.toml}. If \code{NULL} (default), sanitized R
package name is used instead.}

\item{quiet}{Logical indicating whether any progress messages should be
generated or not.}
}
\value{
a logical scalar indicating whether scaffold updating was successful
}
\description{
When a new version of extendr or rextendr is released, this function updates
relevant scaffolding files to the new specification.
}
\details{
This function does not touch any build artifacts or files or folders
generated when vendoring cargo. Cargo.lock and Cargo.toml are also left
unchanged. Only the following files are re-written:
\itemize{
\item \code{src/entrypoint.c}
\item \code{src/Makevars.in}
\item \code{src/Makevars.win.in}
\item \code{cleanup}
\item \code{cleanup.win}
\item \code{src/rust/document.rs}
\item \code{tools/msrv.R}
\item \code{tools/config.R}
\item \code{configure}
\item \code{configure.win}
}

After updating these files, \code{update_scaffold()} will print a message that
explains what to do next to get your package up-to-date with the latest
versions of extendr and rextendr (provided \code{quiet = FALSE}, anyway). That
will typically include handling dependency resolution, updating Cargo.toml
and Cargo.lock, and vendoring crates for CRAN compliance. Usually, this
will be accompanied by a more detailed blog post explaining the update
process.
}
