% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{use_extendr_badge}
\alias{use_extendr_badge}
\title{extendr README badge}
\usage{
use_extendr_badge(path = ".")
}
\arguments{
\item{path}{File path to the package for which to generate wrapper code.}
}
\description{
Add the version of extendr being used by an R package to its README.
}
\details{
Requires \code{usethis} to be available.
}
\examples{
\dontrun{
use_extendr_badge()
}

}
