% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license_note.R
\name{write_license_note}
\alias{write_license_note}
\title{Generate LICENSE.note file.}
\usage{
write_license_note(path = ".", force = TRUE)
}
\arguments{
\item{path}{Path from which package root is looked up.}

\item{force}{Logical indicating whether to re-generating LICENSE.note if LICENSE.note is already exists.}
}
\value{
No return value, called for side effects.
}
\description{
LICENSE.note generated by this function lists dependent Rust packages infomation.
To use this function, the \href{https://crates.io/crates/cargo-license}{cargo-lincense} command must be installed.
}
