% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license_note.R
\name{write_license_note}
\alias{write_license_note}
\title{Generate LICENSE.note file.}
\usage{
write_license_note(path = ".", quiet = FALSE, force = TRUE)
}
\arguments{
\item{path}{character scalar, the R package directory}

\item{quiet}{logical scalar, whether to signal successful writing of
LICENSE.note (default is \code{FALSE})}

\item{force}{logical scalar, whether to regenerate LICENSE.note if
LICENSE.note already exists (default is \code{TRUE})}
}
\value{
text printed to LICENSE.note (invisibly).
}
\description{
LICENSE.note generated by this function contains information about all
recursive dependencies in Rust crate.
}
\examples{
\dontrun{
write_license_note()
}
}
