% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/license_note.R
\name{write_license_note}
\alias{write_license_note}
\title{Generate LICENSE.note file.}
\usage{
write_license_note(path = ".", quiet = FALSE, force = TRUE)
}
\arguments{
\item{path}{Path from which package root is looked up.}

\item{quiet}{Logical indicating whether any progress messages should be
generated or not.}

\item{force}{Logical indicating whether to regenerate LICENSE.note if LICENSE.note already exists.}
}
\value{
No return value, called for side effects.
}
\description{
LICENSE.note generated by this function contains information about Rust crate dependencies.
To use this function, the \href{https://crates.io/crates/cargo-license}{cargo-license} command must be installed.
}
